HT_21/requirements.txt
django-crispy-forms==1.14.0
Pillow==9.0.1
Django==4.0.1
Pygments==2.11.2
djangorestframework==3.13.1
